+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/settings/base.py
@@ -308,7 +308,7 @@ STATICFILES_FINDERS = [
 
 MEDIA_ROOT = join(PROJECT_ROOT, 'media')
 MEDIA_URL = '/media/'
-
+WAGTAILMEDIA_MEDIA_MODEL = 'core.MoloMedia'
 
 # Django compressor settings
 # http://django-compressor.readthedocs.org/en/latest/settings/